Move mspileupError from MSPileup to Data module
